no message
git-svn-id: svn://svn.code.sf.net/p/actiongame/code/trunk/ac@429 e016b1c5-ed68-44aa-a2f7-31cac88cffde

+++ source/src/rendergl.cpp
@@ -243,7 +243,7 @@ VARF(gamma, 30, 100, 300,
     };
 });
 
-void trypitch(int i) { player1->pitch=i; }
+void trypitch(int i) { player1->pitch = (float)i; }
 COMMAND(trypitch, ARG_1INT);
 
 void transplayer()
